rename `focusControl` to `qmlUtils`
