+++ client/utils/qmlUtils.h
